requirements.txt
Django==1.4.2
boto==2.6.0
dj-database-url==0.2.1
django-storages==1.1.5
gunicorn==0.16.1
psycopg2==2.4.5
wsgiref==0.1.2
